Bug fix save&restore logging
