+++ app/save-and-restore/logging/src/main/resources/META-INF/services/org.phoebus.framework.adapter.AdapterFactory
@@ -1,1 +1,2 @@
-org.phoebus.applications.saveandrestore.logging.SaveAndRestoreActionAdapterFactory
+org.phoebus.applications.saveandrestore.logging.SaveSnapshotActionAdapterFactory
+org.phoebus.applications.saveandrestore.logging.RestoreSnapshotActionAdapterFactory
