+++ app/save-and-restore/logging/src/test/java/org/phoebus/applications/saveandrestore/logging/SaveSnapshotActionAdapterFactoryTest.java
@@ -26,31 +26,19 @@ import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class SaveAndRestoreActionAdapterFactoryTest {
+public class SaveSnapshotActionAdapterFactoryTest {
 
-    private SaveAndRestoreActionAdapterFactory saveAndRestoreActionAdapterFactory = new SaveAndRestoreActionAdapterFactory();
+    private SaveSnapshotActionAdapterFactory saveSnapshotActionAdapterFactory = new SaveSnapshotActionAdapterFactory();
 
     @Test
     public void testGetAdaptableObject(){
-        assertTrue(saveAndRestoreActionAdapterFactory.getAdaptableObject().isAssignableFrom(RestoreSnapshotActionInfo.class));
+        assertTrue(saveSnapshotActionAdapterFactory.getAdaptableObject().isAssignableFrom(SaveSnapshotActionInfo.class));
     }
 
     @Test
     public void testGetAdapterList(){
-        List<? extends Class> list = saveAndRestoreActionAdapterFactory.getAdapterList();
+        List<? extends Class> list = saveSnapshotActionAdapterFactory.getAdapterList();
         assertTrue(list.get(0).isAssignableFrom(LogEntry.class));
-    }
-
-    @Test
-    public void testAdaptRestoreAction(){
-        RestoreSnapshotActionInfo restoreSnapshotActionInfo = new RestoreSnapshotActionInfo();
-        restoreSnapshotActionInfo.setSnapshotName("snapshot name");
-        restoreSnapshotActionInfo.setFailedPVs(List.of("pv name"));
-
-        LogEntry logEntry = saveAndRestoreActionAdapterFactory.adapt(restoreSnapshotActionInfo, LogEntry.class).get();
-        assertTrue(logEntry.getTitle().contains("snapshot name"));
-        assertTrue(logEntry.getDescription().contains("snapshot name"));
-        assertTrue(logEntry.getDescription().contains("pv name"));
     }
 
     @Test
@@ -58,7 +46,7 @@ public class SaveAndRestoreActionAdapterFactoryTest {
         SaveSnapshotActionInfo saveSnapshotActionInfo = new SaveSnapshotActionInfo();
         saveSnapshotActionInfo.setSnapshotName("snapshot name");
 
-        LogEntry logEntry = saveAndRestoreActionAdapterFactory.adapt(saveSnapshotActionInfo, LogEntry.class).get();
+        LogEntry logEntry = saveSnapshotActionAdapterFactory.adapt(saveSnapshotActionInfo, LogEntry.class).get();
         assertTrue(logEntry.getTitle().contains("snapshot name"));
         assertTrue(logEntry.getDescription().contains("snapshot name"));
     }
